refactor(#80): 로비 Lua 실행 책임 분리
